consolidate site-admin routes
